+++ examples/sitemap-ext/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @example/sitemap-ext
 
+## 0.1.12
+
+### Patch Changes
+
+- @inox-tools/sitemap-ext@0.2.10
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [b6d23ab]
+  - @inox-tools/sitemap-ext@0.2.9
+
 ## 0.1.10
 
 ### Patch Changes